feat: Working game

+++ src/common/utils.cpp
@@ -51,5 +51,5 @@ int Utils::bitCount(uint8_t number) {
     return static_cast<int>(lookupTable[number]);
 }
 uint64_t Utils::micros() {
-    return std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
+    return static_cast<uint64_t>(std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::system_clock::now().time_since_epoch()).count());
 }
